refactor BasicActivityInfo and add frequency to goal description
