+++ src/util.js
@@ -84,3 +84,43 @@ export function extractActivityLists(state, day){
 export function isToday(date){
   return date?.startOf('day').toISO()==DateTime.now().startOf('day').toISO()
 }
+
+export function frequency(activity){
+  let frequency 
+    switch(activity.repeatMode){
+      case 'weekly':
+        if(activity.goal=='check'){
+          frequency = `${activity.timesPerWeek} days per week.`
+        }else{
+          const expression = getPreferedExpression(activity.timeGoal)
+          frequency = `${expression.value} ${expression.unit} per week.`
+        }
+        break
+      case 'select':
+        let days = ''
+        const labels = {1: 'Mon', 2: 'Tue', 3: 'Wed', 4: 'Thu', 5: 'Fri', 6: 'Sat', 7: 'Sun'}
+        for (let day in activity.weekDays){
+          if (activity.weekDays[day]){
+            days = `${days} ${labels[day]}`
+          }
+        }
+        if(activity.goal=='check'){
+          frequency = `Do on ${days}`
+        }else{
+          const expression = getPreferedExpression(activity.timeGoal)
+          frequency = `${expression.value} ${expression.unit} on ${days}`
+        }
+        break
+      case 'daily':
+        if(activity.goal=='check'){
+          frequency = "Every day."
+        }else{
+          const expression = getPreferedExpression(activity.timeGoal)
+          frequency = `${expression.value} ${expression.unit} every day.`
+        }
+        break
+      default:
+        frequency = 'ERROR'
+    }
+  return (frequency)
+}
